Spanner schema SDL file for computations tables
It it one table with two interleaved children.

Computations
 |
 ` ComputationStages // Provides a history
   |
   ` ComputationBlobReferences // Provides references to BLOBs

The Computations table keeps track of a computation at a worker
which is its current stage, if it is being worked on, where to
send data etc.

The ComputationStage table is a child table to provide history of
when changes occured with the computation which will help us build
metrics for comptation stages.

The ComputationBlobReferencesTable is a list of BLOBs used
in the stage. It is a child table because there can be multiple inputs
or outputs to single stage (like receiving a sketch from all other workers).

The schema follows the recomendations of OM Relational Database Columns
http://docs/document/d/1urOaexh_8L8DZsEr-gHKTjBEOxQp2QiCx3hvs4SBgBI
which means the columns that are not included in an index or are a
timestamp are put into a proto column. However, public spanner doesn't
offer protobuf columns so they are stored as both serialized bytes, and
as JSON (for ad-hoc query support).

Change-Id: Ieec0d10d8190553e67331c2c815bf8213ded5bef