ProtocolStateEnumHelper interface for managing state enum(s)
This interface privides validation checks for what states are
allowed to follow others as well as a way to convert from an
enum to a number for storing in a database.

You may think "hold on isn't there ordinals for enums?" Well yes,
but going from an ordinal to a kotlin enum in a generic is not
possible without using some unsafe reflection calls. Also raw
enums would not be able to take advantage of protobuf's UNRECOGNIZED
distinctions

Change-Id: Ia980c4cd04734b48e2f718a6b1c03784d3045b83