Implement claimTask()
While doing this I realized that allow_commit_timestamp forces timestamps
to be in the past, which goes against the way LockExpirationTime is
intended to be used as locks expire in the future after a task is
claimed. Thus that option was removed from the field. The GcpSpannerComputationsDb
object now takes a Clock for testin purposes. This makes it possible to
test the timestamp fields in database.

Change-Id: I5f8b739f785e66f49ffd00ea2cea324325f0eaf5