Make GlobalComputationId unqiue in Computations table
I was allowing for a one-to-many mapping of computation ids from
the kingdom to local compuation ids in the duchies, but this
is an overly complicated model that likely would be hard to query
and work with. If a computation gets killed by the kingdom and
restarted later, it would make sense for the kingdom to know that
mapping of ids which realted to the same desired outputs.

Change-Id: I5f36914e352147e5da40fb8616d18fdae207e085