Write an Attempt row iff an attempt was started
ComputationStageAttempts.BeginTime can no longer be NULL.

This makes the state of the database a little more understandable
in that the existance of a row says if an attempt was made or not.
Before queries would have to rely on the begin timestamp being set to null.

Attempts are written when a lock is aquired on a task or if the lock
is retained on a stage transitions which is logically the same as
inserting a new stage then acquiring the lock on it.

When a lock is acquired on a task with an expired lock, the ongoing
attempt is ended and a new row is written.

Change-Id: Ie69de29bb2d1d6434b8b29ae775ad8c2e48c5391