Reconfigure ComputationMutations to have type saftey for stage values
The top-level functions were moved into a class that accepts the
type of the stage and the type of the stage details being stored
in the database. This type information was propogated all the way
up to the ComputationManager. The type information allows mutations
to be created from the right enum and not an arbitrary Long value.

This change also adds a helper class to write stage specific details
for a computation stage.

Change-Id: I813beef5059795c227ecfae9c719cd0778b670ce