Implement ComputationsBlobDatabase in Google Cloud Storage
The GcpStorageComputationsDb supports writing, reading, and deleting
of blobs from blob store, as well as a function for getting a name
which can be used to write a new blob for a computation stage. This
name has some randomness so there is not write contention if a peasant
lost the lock.

Change-Id: I61c5e31072d097239867ca96f10506221f8d6755