Simplify ComputationManager transition state interface
The function moves a computation from one stage to another while
writing new blobs, keeping references to old blobs, and making note
of needed output blobs. All of which required the caller to provide
a BlobRef e.g., a pair of blob id and a path to the blob. However the
blob id is not actually needed for the transition and is in fact
ignored because it is the reference from the previous stage,
not the new one.

The new function interface takes Strings, which is just the path
portion of the BlobRef, for the blobs to write and the blobs to carry
forward. The blobs required for output has been changed to an Int
as the only thing that matters when changing state is the count of
expected output blobs.

Lastly, a check was put in place to make sure all output blobs
references in the database have a path written (all blobs that are
outputs of the current stage have been written) before changing the
state.

Change-Id: Ia01d5d26ac9a710a720d8c384100ab15a1b7b93f