Add SketchAggregationComputationManager
This is a computation manager specific to the joining sketches in an
MPC. It provides the right generic typing for the computation and a
function for consistently calling transitionState based on the stage
being entered, which should greatly simplify the duchy gRPC service and
the Peasant jobs.

There is also a function added to create a
SketchAggregationComputationManager running on GCP specific
infrastructure (Spanner and Google Cloud Storage). A couple of tests
were added to test the correct path through the protocol for both the
primary and secondary duchies.

Change-Id: Iecbf72292a675cb1e801b753f05860234cbd2447