Remove blobsToWrite from ComputationManager.transitionState()
In short it isn't really needed.

When I first designed the interface I thought it would be a useful way to move
from one state to another, and specifically with WAIT_FOR_SOMETHING states in
mind. My thought was a gRPC handler would get a message and transition into a
new state while writing the received message in the same function call. This
paradigm wouldn't work to well for WAIT_SKETCHES because we also need to track
who's sent a sketch. I think its nice if an output blob is
written as part of WAIT_CONCATENATED, WAIT_JOINED via receiving the desired
method. One it clearly shows that the wait for some needed data and two it
allows us to ensure the a reference to the output is written before leaving
the state.

Removing this paramater makes the function have one parameter for inputs
and one for outputs. This allows for easier to understand names and
behavior for the caller to the function. It also means the function can
be much more simple. It doesn't have to combine two sources of input
blobs.

Change-Id: I1f507e1ac8e61152a0701a29c29d82f7035977af