Refactor states to remove the RECEIVED_* states.
The RECEIVED states were a way to pass the computation
from worked on by a gRPC server to a Farmer. This
complicated the stage graph with sentinel states that
only kept track of an output from the proceeding WAIT_
step.

With this refactor the gRPC server receives data, writes
it as output and then moves to the TO_DO_A_THING stage.
The Farmers operate on the TO_DO_A_THING stages, and the
attempts of them are the times the farmers are executed
on them.

With this change a second attempt on a failed stage lookes
the same as an attempt of the stage the first time, which
is to write a new attempt row to the stage.

https://docs.google.com/document/d/1JTjK42BEZpTZDMNPvHb1xbmRUSVHcM1hjEmI9AiUpwg

Change-Id: Ic698185268fb824686841ee90400afe2ca94dae8