SketchAggregationComputationManager uses the storage service
It is no longer an implementation of the ComputationManager class
but is still the place which encapsulates the logic specific to
the LiquidLegions sketch aggregation protocol.

This change also replaces ComputationToken kotlin class with the
proto version in FakeComputationStorage. As a result,
the FakeComputationsRelationalDb is no longer needed and was folded into
the FakeComputationStorage class.

Change-Id: I3ca4648c8d7989fefbca8dc5c59d96062463180a