Refactor newBlobPath method out of ComputationsBlobDb.
ComputationsBlobDb is an interface that abstracts away the details of the blob storage system used underneath. Naming blobs is an implementation detail of the ComputationManager as you can see from the fact that the newBlobPath method expects a ComputationToken. Also appending a random suffix seems like something we would want to do in general not just for GCS. Refactoring naming up into the ComputationManager keeps ComputationsBlobDb generic and should make it easier to reuse the interface in other parts of the Open Measurement system and to add support for other blob storage systems later.

One reason to consider keeping newBlobPath in the ComputationsBlobDb interface is to handle variation in naming restrictions across different blob storage systems. However forward slash separated strings seem to be the standard for GCS, S3 and Azure Storage so this does not seem like a likely issue to encounter.

Change-Id: Id45f96a03a5945e023a66ed9babf360f16df343f