ReadOnlyComputationsRelationalDb interface and spanner implementation
There were two notions of a computation token, one used when making
edits to the database and another returned by the
ComputationStorageService API that has information needed to perform
steps of a computation. These two tokens were conflated in
ComputationsRelationalDb class. That class now only uses the token
that is useful for changing data; now named
ComputationStorageEditToken. The class no longer knows anything about
the ComputationToken used in the API, but a new interface the
ReadOnlyComputationsRelationalDb knows how to make such tokens from data read
from the database. This distiction means less data needs to be packed
into tokens being used in database transactions.

There are a couple of changes to the ComputationsRelationalDb of note
here too.

The blobs where added to the proto version a ComputationToken
so the caller doesn't have to make additional calls to find
the object keys, when they will most certainly need them.
This makes the readInputBlobs function obsolete in
ComputationsRelationalDb. So it was removed.

The Stage specific details were also moved into the ComputationToken
so likewise the readStageSpecific details token was removed.

The readGlobalComputationIds function was moved into the new
ReadOnlyComputationsRelationalDb interface.

Change-Id: I4f34e277026f886f4e3cb026bcd7dd1ac1e6cc60