Implement the Computation Storage Service
The service was almost fully implemeted in the
FakeComputationStorageService. The fake was extended
to use an implementation of database, and the fake database
now extends that interface. This means the fake service is
now the same as the real service, but backed with a fake database
and not custom rpc logic.

Change-Id: I02111afab4f3dac172446eecee2b5f7e444f1472