Refactor SketchAggregation helpers into single object
LiquidLegionsSketchAggregationProtocol is a top level ojbect containing
the class/onject definitions of helpers specific to the liquid legions
protocol

LiquidLegionsSketchAggregationProtocol.EnumStages
LiquidLegionsSketchAggregationProtocol.EnumStages.Details
are for working with the enum values directly

LiquidLegionsSketchAggregationProtocol.ComputationStages
LiquidLegionsSketchAggregationProtocol.ComputationStages.Details
are for working with the protos wrapped in the ComputationStage
proto. This second grouping will be the most common use as they
are the API level.

Change-Id: I3e819d894590c8d78cde1a0e33fcc806183dc72d