Add a MetricValueDatabase interface with a Google Cloud Spanner implementation.
This is following the principle of ensuring that internal IDs do not leak outside of the class which implements the database interface. Therefore if we want to allow for cross-references between computation and metric value Duchy Spanner tables, we'll need to have a single class for all Spanner interaction in a Duchy.

Bug: 159036541
Change-Id: I0997be5dbfab44af281fc13d74f02ca630b2d92d