Add the TO_CONFIRM_REQUISITIONS and WAIT_TO_START stages
1. The herald creates a new computation at stage TO_CONFIRM_REQUISTION,
the list of requisitions are stored in the ComputationStageDetails
2. A mill would claim the job and fetch all requisitions and combine
them together and store the result in the local blob store for future
usage.
3. After confirmation, non-primary duchy enters WAIT_TO_START and wait
for its herald to move it further to TO_ADD_NOISE, primary duchy enters
WAIT_SKETCHES and wait for all non-primaries to send their sketches to
it.

Change-Id: I5869a8706b8c670dfc09716d2b07fb985f027874