Split the db/duchy package by domain object
The duchy has two types of tables and databases abstractions in the
duchy database. One for uploaded publisher data (metric values) and one
for managing computations. Recently we've decided to keep storage
separate from one another. To make this clearer and reflected in
the code, there are now two sub-packages to db/duchy. They are
metricvalue for uploaded metric values via the publisher data service
which are stored at the duchy; and computation for computation management
like things like interactions with the computation database tables
and reading/writing the intermediate results.

Change-Id: I9430201cba4755d50e5057143e66bcda419bc843