Adds DuchyId to Requisition proto and in DB.
Requisitions are fulfilled by Duchies. In order to keep track of which
Duchy has fulfilled which requisitions, we add a column in Spanner to
the Requisitions table. This is exposed in the Requisition proto and set
in the FulfillRequisition RPC methods.

This also determines how services will get access to duchies'
authenticated identities: see DuchyAuth.kt for this. Note this doesn't
actually do any authentication or authorization, it just defines a
pattern for communicating their verified identity.

Change-Id: I4cd9c347026809b3dd05bb36c9fc25f9dbab8f2b