Add utilities for passing duchy id in gRPC side channel.
This CL doesn't handle actual authentication or authorization,
but it provides a mechanism for passing the Duchy ID in the side channel.

This CL takes no stance on how authz/authn should be performed. We
simply use the side channel to pass the duchy id as an alternative to
putting it in the request body because:

- This is a uniform, automatic way to transmit the identity.
- Duchies can configure this in one line of code with little opportunity
  for error.
- It avoids polluting the API with extra duchy_id fields.

If, eventually, authz is done through gRPC's credential mechanisms, it
may require only small changes for this identifier to be secure.

Change-Id: Ic80c9821a84159c40886115552b6d0cf46f34d0f