Check status of computation on first message in stream
There are cases, like when the blob was already written that
we just need to ack the message, but don't need to collect
the whole sketch. This change makes each of the handle messages
in the computation control service retrieve the token for a
computation when processing the first item in the stream. Then
that can be used to see if the rest of the stream needs to be
collected.

In this change the bytes in the flows will not be collected
if the blob was already written, i.e. the output blob ref path
is not empty, or if the computation has already transitioned
to the next stage. In both cases, the control service acks
the message so the sender can update the local stage of the
computation.

Change-Id: I1a24729cc16b6b5c75e8ae5eede773279a32cf48