A first pass at yaml file for all images in kingdom and one duchy
This is rife with TODOs. I created this file with by scraping help
from all the *_image.binary files. Flags are set to the default value
if one exist from the picocli description, or are set to a TODO message.
flags that end with -target are set to the
K8_SERVICE_HOST:K8_SERVICE_PORT, if that name was obvious from the bazel
build target. Otherwise, it also is set to a TODO value.

I also added an option to deployToKind to deploy this still new yaml
file. Enenvutally they should be recombined.

Change-Id: I39644108f1c1ff85faac6d06f5ec1030a52a1d4c