Don't use logAndSuppressExceptionSuspend in the mill
It is causing timeout in the integrationTest. All codes in the test have
been executed but the test never ends. There might be an issue during
the test clean up, for example, the mill could not be killded/

Since we are already catching all Expections in
processNextComputation(), removing logAndSuppressExcetion is OK

Bug: 168618979
Change-Id: I44663784bf835e3e343098685a25e4bb20d1fc61