Add a delay_second field to the EnqueueComputationRequest.
Currently EnqueueComputation is not used anywhere except in tests.
After adding this field, the mill can call this method to release the
lock if a transient error happens during the work. This allows the next
mill to pick up the computation sooner without needing to wait for 5
minutes (current default lock expiration time)

Change-Id: I4c19c3c7b27a8d228995da609ddf6c4aa8d24d92