Remove parallel ARRAY_AGG from ComputationTokenProtoQuery.
ARRAY_AGG has nondeterministic order, so it can't used for selecting
parallel arrays. Instead, this switches the query to use a subquery to
build an array of structs. This correctly handles empty arrays, unlike
ARRAY_AGG, which will make a single item of null values.

This adds a test case to for the case where there are no BlobReferences.

Change-Id: I163956c23f6e42f754c66abb7d35a72f96adf11d