Ensure Spanner objects from SpannerFromFlags are closed.
The Spanner Java API documentation indicates that it is important to ensure that the Spanner objects are closed when they are no longer in use, for example on process shutdown.

This also changes the startup behavior of binaries that use SpannerFromFlags. These will attempt once to ensure the Spanner DB connection is ready with an exception on timeout. Due to the fact that closing Spanner can take on the order of a minute when no connection is established, the timeout error is immediately logged. The ready check is also updated to use the asynchronous API to avoid observed blocking.

Change-Id: Ia52f4242f76c87101c6ac0475ae331e2bdce38f4