Add BlobMetadata type PASS_THROUGH
There are some blobs that inputs to a stage which are also
carried over as output of the stage. There was no representation
of that in the DB so it had to be encoded in kotlin logic. This
makes it harder to have a generic handeler of messages in the
computation control service.

Change-Id: I5cd39af2f7b3610ab8e5babbcc4c9d90bc5318df