Add an UpdateComputationDetails method to the ComputationsService.
There will be 2 usage cases so far.
1. The LL_V2 protocol estimates the reach in round 1, the result will be cached in the computation_details and attached to the final report in round 3 when the mill reports to the kingdom.
2. The CompleteFrequencyPhase result(freqeuncy distribution map) would be cached inside the computation_details instead of in the blobstore. The cache will be used for retry if necessary.

Note that, most fields in the computation_details are not suitable for updating. e.g., the computationRole, the blob_storage_prefix, etc. However, since the service is an internal only service, we assume the caller knows what it is doing and won't mess up it is own computation.

Change-Id: I45759563bb77c672025f45f5d673c4a433d62412