Drop dependencies on @debian_java_base, adding constraints for Distroless Java. (world-federation-of-advertisers/experimental#72)
This requires that java_image targets that depend on JNI for protocol encryption be built on a platform that has an appropriate version of glibc. The host platform can be specified using the `--host_platform` option to the Bazel CLI. For example, Ubuntu 18.04 "bionic" can be specified using `--host_platform=//build/platforms:ubuntu_18_04`.

If using the included RBE or Bazel container configuration, this will already be specified correctly.