Add Kingdom-internal Panel Match data-layer gRPC service definitions.
For now, this adds a minimal set of methods required to implement most
of the functionality of the public facing APIs.

As more functionality is required, these will need to be expanded.
Building them incrementally has the advantage of not needing to
implement methods that are later determined to be unnecessary.

See world-federation-of-advertisers/cross-media-measurement#3.