Split apart KingdomRelationalDatabase. (#19)
This moves functionality into super-interfaces of
KingdomRelationalDatabase. This also splits out test-only logic into a
test-only database helper class.

The next steps are:

1. Split apart implementations of KingdomRelationalDatabase into
implementations of the super-interface.

2. Transition clients of KingdomRelationalDatabase to clients of the
super-interfaces.

3. Delete KingdomRelationalDatabase.

These changes are important because, as part of #3, a fair amount of
new database functionality is needed. However, most clients only need a
tiny subset of the exposed interface. By splitting it up, we are able to
logically isolate functionality and narrow the public interfaces seen by
most clients.