Update LLV2 protos to fit the per computation ElGamalKey design.
1. two extra stages are added to the protocl INITIALIZATION_PHASE and
WAIT_REQUISITIONS_AND_KEY_SET.
2. In CONFIRM_REQUISITIONS_PHASE, mill will only confirms that local
requisitions are present, it won't read and combine the actual
requisitions and write the result as an output anymore. The actual
requsitions are read later in the Setup phase. This saves storing
another copy of all requistions.
3. ToConfirmRequisitionsStageDetails proto will be deprecated. The
requisitionList would be written inside the computationDetails instead,
since it is used in multiple stages.