Update LLV2 protos to fit the per computation ElGamalKey design.
1. Two extra stages are added to the protocl, i.e., INITIALIZATION_PHASE and
WAIT_REQUISITIONS_AND_KEY_SET.

2. CONFIRM_REQUISITIONS_PHASE is changed to CONFIRMATION_PHASE. In this
stage, the mill will confirms that
-- all local requisitions are present (it won't read and combine the
actual requisitions and write the result as an output anymore. The actual
requsitions are read later in the Setup phase. This saves storing
another copy of all requistions.)
-- EDP signatures and other duchies signatures are all valid.

3. ToConfirmRequisitionsStageDetails proto will be deprecated. The
requisitionList would be included in the serialized KingdomComputation
in the ComputationDetails.