Update the confirmationPhase processor in the llv2 mill.
Behavior change:
previously, the mill would fetch all local requisitions from the
metricValueService and store the combined sketch locally for the future
processing.
Now, the mill just verify requistions status based on the database information,
no blob read/write is done.

so the output blob number of the INITIALIZATION_PHASE changes from 1 to
0. And there is no input blob for the WAIT_TO_START and
WAIT_SETUP_PHASE_INPUT stages any more.

The combinedPublicKey and particallyCombinedPublicKey are also
computated and stored in the computationDetail in this phase.