Change ExchangeStepAttemptDetails inputs/outputs
Changes the ExchangeStepAttemptDetails proto to have a shared inputs and
shared outputs field. Changes those fields from bytes (blob) to string
with the expectation that the storage step will pass signed cloud
storage urls containing the exchange data blob rather than the actual blob
itself. These blobs would likely be quite large and not work as part of
an API reponse directly. As this proto is not yet used it does not seem
necessary to reserve the existing shared_ouputs and use a new field
name despite the change in field type.