Update task execution to lazily stream inputs and outputs (world-federation-of-advertisers/panel-exchange-client#75)
* Update task execution to use a stream abstraction

Uses StorageClient.Blob more directly and adds a method to convert blob
data to bytestring.

Implements an InMemoryStorageClient to make it possible to test.

The goal is to avoid reading in the entirety of an input (or collecting
an entire output before writing) all at once to avoid memory issues.
While the existing exchange task steps can and do collect their inputs,
these will be revisited in a later PR and other tasks with larger inputs
and outputs will leverage these streams.