Refactor Cryptor. (world-federation-of-advertisers/panel-exchange-client#91)
(1) This REMOVES the batch methods Cryptor -- they add unnecessary
complexity. One reason they originally existed was to make the class
threadsafe but that's not needed. Instead, it's explicitly documented as
not threadsafe.

(2) Renames Cryptor to DeterministicCommutativeCipher to be more
specific.

(3) Misc code clean-ups, such as removing assumptions about error
messages.