Better support containerized builds when developing on Linux environments.
This changes the remote Bazel output base to be $PWD/bazel-container-output, allowing linux-x64 hosts to run binaries built inside the container by creating a corresponding symlink pointing to the volume mount point.

This also mounts the gcloud SDK config in Bazel container so that the container an access the host machine's credentials for GKE deployment.