Fix the GKE correctnessTest deployment.
The test was broken since we have made plenty changes to the system.
For example, principal authority and mark variables related changes.

The PR also
1. uploaded a synthetic-labelled-events.csv for completing correctnessTest
on GKE
2. switched the order of EDP and MC creation in the resourceSetupJob to
make sure we call the kingdom public API first. Otherwise, the job may
call the internal kingdom API successfully and fail later when calling
the public API. If this happens, the retry would fail constantly since
the EDPs are already created.
3. Deleted the labels in the NetworkPolicy, since NetworkPolicy doesn't
supoort labels.