Never restart Kingdom resource setup job. (#449)
The individual resource creation is not repeatable due to certificate uniqueness, so if there's a partial failure then the records need to be wiped from the DB before it can be run again.

This also fixes a bug where the resource setup job's dependencies were not being respected (i.e. it was not waiting for the Kingdom internal server to be running).