Ensure an exchange step is not its own prerequisite.
In general, exchange steps should not have the same blob key in both their input map and output map, as this causes the step to be its own prerequisite and thus can never enter a READY state. The only exception is CopyFromPreviousExchangeStep which must have the same input and output blob key in order to work correctly. For this type of step, explicitly remove its index from the prerequisite indices so that it can be properly scheduled.