Catch more RPC exceptions.
StatusException should be caught at the lowest level, i.e. the call site of the RPC. This is because how a given error should be handled depends on which method was being called. If it needs to be handled at a higher level, it should be wrapped in another exception type to make that explicit to calling code.