Use kustomization rules for local environment secret generation.
This drops the need to manually create a secret as a separate step and passing its generated ID to build targets. It only covers the local environment at the moment. Future work could include extending it to the dev environment, as well as updating guides on how to use Kustomize directly for deployment.