Add Bazel test target for CMMS KinD correctness test.
The test is implemented using JUnit and assumes that KinD is installed and in the test PATH. It uses the Kubernetes API to interact with the cluster after creation. The test methods are run locally using port forwarding rather than deploying the test harness (FrontendSimular) to the cluster. Resource setup is also run locally.