Implement logic to update Data Provider duchy list (#934)
Create ReplaceDataProviderRequiredDuchies class to update duchy list in
DataProviderRequiredDuchies table. The return value of the
TransactionScope.runTransaction method in this class is the DataProvider
object with the updated duchy list

Implement replaceDataProviderRequiredDuchies in
SpannerDataProvidersService to call the new
ReplaceDataProviderRequiredDuchies class with the
ReplaceDataProviderRequiredDuchiesRequest. This method will throw an
error if the externalDataProviderId is not present in the request.

Create test that checks:
1. the returned DataProvider has the updated duchy list
2. the DataProviderRequiredDuchies table is updated


https://rally1.rallydev.com/#/604612930531d/iterationstatus?detail=%2Ftask%2F697342533129