Derive metric states (#963)
Motivation:

* In the current design, the metric state is updated using a on-demand
strategy -- only updated when the public `get` or `list` is called. This
causes the internal metric state is not consistent with what the states
of the associated measurements indicate. For example, a metric might
have all measurements with SUCCEEDED states, but itself is PENDING.



Design change:

* The metric state is essentially a derivative from the states of the
associated measurements. Therefore, we could instead determine the
metric state in the public server call and remove the state column in
the internal metrics DB.

* Similarly, the metric result can be calculated easily on the public
side, so we don't need to store metric result in the database.