Add MultithreadingHelper to refactor multithreaded LLV2 (#976)
## Background
To multithread the execution of LLV2. Currently all for-loops are
single-threaded but iterations are actually independent. They can be
refactored into a multi-threading pattern.
Most operations rely on
[ProtocolCryptor](https://sourcegraph.com/github.com/world-federation-of-advertisers/cross-media-measurement/-/blob/src/main/cc/wfa/measurement/common/crypto/protocol_cryptor.cc).
Unfortunately, the class ProtocolCryptor is not thread-safe. Thus,
multiple identical ProtocolCryptors must be created and each thread
should have a dedicated ProtocolCryptor.
Current iterations append results directly to the output variable. This
is not a good pattern for multiple threading if the order matters
because the execution order of iterations is not deterministic.

## Design 
The general idea is to have a helper class assigning tasks to multiple
threads and collect output after execution.
### Parameters
The helper has two primary parameters: # iterations, and a
lambda expression.
The lambda expression reuses current llv2 code as the task to run. The helper execute the lambda expression with index
from 0 to (# iterations -1). 
Capture list of the lambda expression has all the context to execute
iterations. No major change required from current for-loop blocks.
### Results
Lambda expression is responsible for handling results from iterations
and making sure it is thread-safe.
### Cryptors
The ProtocolCryptor is not thread-safe. Each thread should have its own
ProtocolCryptor to execute lambda expressions.
The helper can be reused multiple times with different types of input
and output. ProtocolCryptor only needs to be initialized once. For
example SetupPhase, it has multiple for-loops to add various types of
noise. Only one helper need to create and reuse.
Some procedure does not require ProtocolCryptor thus can skip
`setupCryptors`.

## Class Interface
See the header file.

## Usage
See the unit test.