adding grpc-gateway to generate bff reporting definitions (#1006)
This is to create the proto messages for gRPC-Gateway. gRPC-Gateway uses
a Go gateway, so these messages will be converted to Go. The Kotlin
messages will be added later when setting up the gRPC service (not the
gateway).

Right now the BFF proto messages are exactly the same as the Reporting
service v2alpha messages. This will surely change, but we're still
working out how the UI will look and not sure how to make those messages
yet. So for now, just going to use the existing proto definitions to put
something here.