Fix that only allows creation of ModelRollouts with RolloutPeriodStartTime after that of previous ModelRollout (#1076)
Update query in readModelRolloutData to return the ModelRollout for a
given ModelLine with the latest RolloutPeriodStartTime in
CreateModelRollout.kt. Update CreateModelRollout.kt to throw
ModelRolloutInvalidArgsException if the RolloutPeriodStartTime of the
ModelRollout to be created precedes that of the last ModelRollout.
Create tests in ModelRolloutServiceTest.kt. Update 'createModelRollout
succeeds with multiple model releases' test to adhere to new
RolloutPeriodStartTime limitations.