Update EDP simulator to stop writing plaintext sketches to storage. (#1112)
This means that the MC simulator cannot read these sketches to compute expected results. It is updated to compute them using an EventQuery implementation, which must be the same one used by EDP simulators in that environment.

RandomEventQuery is dropped since it is incompatible with this change. Its usages are replaced with SyntheticGeneratorEventQuery.