Circulate direct computation methodology from protocol config to measurement result. (#1154)
In this PR, the implementation of the [API changes in
CMM-API](https://github.com/world-federation-of-advertisers/cross-media-measurement-api/pull/163)
is done. The main changes are
1. new direct protocol config,
2. additional info in measurement results,
3. new enum constants in `ProtocolConfig.NoiseMechanism`,
4. EDP simulator choose the noise mechanism based on the overlap of its
preferences and the noise mechanism options in the direct protocol
config.

Any future measurement creations will contain certain direct protocol
configs based on the measurement types. For the backward compatibility
for existing measurements and requisitions, the empty internal direct
protocol configs will be converted with the options that would have been
valid at the time the Measurement was created.

EDPs are restricted to use continuous noise mechanisms by setting the
validation in `V2alphaPublicApiServer`. Discrete Gaussian noise will be supported in the future.