Store set operation compilation IDs (#1100)
Part of the changes for implementing error bars in reporting server. To
easily find the union of any two weighted measurements in a metric, we
store the binary representation of each measurement from the binary
representation of the weighed subset union created during set expression
compilation of a reporting set.