Store measurement computation details in reporting for error bars (#1184)
In addition to storing results in the internal measurements, we also
store the computation details about the measurement creation. The
structure of the internal measurement result is now same as the one in
Kingdom -- measurement results are in a list to represent results from
different data providers; noise mechanisms and computation methodologies
are added to `Result`. Note that unlike the Kingdom `Result` which
stores only direct noise mechanisms and direct methodologies, here we
store both direct and MPC noise mechanisms and methodologies.

The additional computation details can be used to compute measurement
variances and potentially used to identify measurements. No DB change is
needed as the changes happened in `Details`.