Restrict metric set operations based on metric types (#1247)
## TL;DR
This pull request introduces a validation check for frequency histogram
metrics to ensure they are only computed on union-only set expressions.
It also updates the unit test name of when a reporting set is not
accessible to the caller and adds a test case for the new validation
check.

## What changed
1. In `MetricsService.kt`, a validation check was added to ensure that
frequency histogram metrics are only computed on union-only set
expressions. If this condition is not met, an `INVALID_ARGUMENT` error
is thrown.
2. The name of the test case `createMetric throws INVALID_ARGUMENT when
reporting set is not accessible to caller` was updated to `createMetric
throws PERMISSION_DENIED when reporting set is not accessible to
caller`.
3. A new test case `createMetric throws INVALID_ARGUMENT when Frequency
Histogram metric is computed on non-union-only set expression` was added
in `MetricsServiceTest.kt`.
4. Comments were added in `metric.proto` to provide more context about
the conditions under which univariate statistics are computed.
 
## Why make this change
This change ensures that frequency histogram metrics are computed
correctly and provides more informative error messages to the user. It
also improves the clarity of the code by adding explanatory comments.