Require maximum_frequency for REACH_AND_FREQUENCY Measurements to be > 1.
This removes a workaround that is no longer needed now that the REACH Measurement type has been released.

Note that this only affects the REACH_AND_FREQUENCY Measurement type. The R/F LLv2 protocol can still be used with a max frequency of 1 for the REACH Measurement type in lieu of the reach-only LLv2 protocol.