Add metric variances to metric results in reporting v2 (#1266)
Compute metric variances in a report using the `Variances` library. Any
insufficiency of inputs to the variance calculation will not set
`univariate_statistics`.

* Reach: compute variances from weighted reach measurements.
* Frequency histogram: compute variances from weighted reach and
frequency measurements.
* Impression count: compute variances only when the operation is
union-only. The variance of an impression metric is the sum of the
weighted impression results.
* Watch duration: compute variances only when the operation is
union-only. The variance of a duration metric is the sum of the weighted
duration results.