Fix the bug that outputs 0 reach if the execution phase crashes right after writing to the output blob. (#1290)
In the execution phase of reach only protocols, the aggregator first
initializes the reach to 0. It then runs an existingOutputOr block to
compute the next token. If the output blob is empty, it updates the
reach and writes stuff to the output blob. Otherwise, it skips the
block, reuses the cache results and sends the reach to the kingdom.

There is a problem when the output blob exists, but the aggregator
crashes before sending the reach to the kingdom. As the existingOutputOr
block is skipped, the reach will not be updated and will stay at the
initialized value (i.e. 0). As a result, the kingdom will receive 0.